+++ app/widgets/MultiSelect.js
@@ -3,9 +3,10 @@ import { View, Image } from "react-native";
 import PropTypes from "prop-types";
 import * as R from "ramda";
 import { ListItem, Text, Icon } from 'native-base';
-import { CheckBox, Tooltip } from 'react-native-elements';
+import { CheckBox } from 'react-native-elements';
 import { getURL } from "../services/helper";
 import { colors } from "../themes/colors";
+import { TooltipBox } from './TooltipBox';
 
 export class MultiSelect extends Component {
   static isValid(value = [], { minValue = 1, maxValue = Infinity }) {
@@ -43,13 +44,9 @@ export class MultiSelect extends Component {
           >
             <View style={{ width: '8%' }}>
               {item.description ? (
-                <Tooltip
-                  popover={
-                    <Text>{item.description}</Text>
-                  }
-                >
+                <TooltipBox text={item.description}>
                   <Icon type="FontAwesome" name="question-circle" style={{color: '#016fbe', fontSize: 24, marginHorizontal: 0}} />
-                </Tooltip>
+                </TooltipBox>
               ) : (
                 <View />
               )}